scripts into one folder + home script implemented
