+++ frontend/gameplay_join/index.html
@@ -19,6 +19,6 @@
             <button type="submit">Send</button>
         </form>
 
-        <script href="script.js"></script>
+        <script type="module" src="../scripts/gameplay_join.js"></script>
     </body>
 </html>
